Fix Google auth missing config handling
