+++ frontend/ui/src/app/auth/sign-in/sign-in-client.tsx
@@ -1,0 +1,191 @@
+"use client";
+
+import { Suspense, useState } from "react";
+import { Eye, EyeOff } from "lucide-react";
+import { authClient } from "@/lib/auth-client";
+import {
+  GOOGLE_AUTH_NOT_CONFIGURED_MESSAGE,
+  GOOGLE_AUTH_START_FAILED_MESSAGE,
+} from "@/lib/auth-errors";
+import { useRouter, useSearchParams } from "next/navigation";
+import Link from "next/link";
+import { useForm } from "react-hook-form";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { z } from "zod/v3";
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
+import { Logo } from "@/components/Logo";
+
+const signInSchema = z.object({
+  email: z.string().email("Invalid email"),
+  password: z.string().min(1, "Password is required"),
+});
+
+type SignInForm = z.infer<typeof signInSchema>;
+
+type SignInContentProps = {
+  googleAuthConfigured: boolean;
+};
+
+function SignInContent({ googleAuthConfigured }: SignInContentProps) {
+  const router = useRouter();
+  const searchParams = useSearchParams();
+  const callbackUrl = searchParams.get("callbackUrl") || "/";
+  const [error, setError] = useState<string | null>(null);
+  const [isLoading, setIsLoading] = useState(false);
+  const [showPassword, setShowPassword] = useState(false);
+  const [isGoogleLoading, setIsGoogleLoading] = useState(false);
+
+  const {
+    register,
+    handleSubmit,
+    formState: { errors },
+  } = useForm<SignInForm>({
+    resolver: zodResolver(signInSchema),
+    defaultValues: { email: "", password: "" },
+  });
+
+  async function onSubmit(data: SignInForm) {
+    setIsLoading(true);
+    setError(null);
+
+    try {
+      const { error } = await authClient.signIn.email({
+        email: data.email,
+        password: data.password,
+      });
+      if (error) {
+        setError(error.message || "Invalid credentials");
+      } else {
+        router.push(callbackUrl);
+        router.refresh();
+      }
+    } catch {
+      setError("An unexpected error occurred");
+    } finally {
+      setIsLoading(false);
+    }
+  }
+
+  async function handleGoogleSignIn() {
+    setError(null);
+
+    if (!googleAuthConfigured) {
+      setError(GOOGLE_AUTH_NOT_CONFIGURED_MESSAGE);
+      return;
+    }
+
+    setIsGoogleLoading(true);
+
+    try {
+      const { error } = await authClient.signIn.social({
+        provider: "google",
+        callbackURL: callbackUrl,
+      });
+
+      if (error) {
+        setError(error.message || GOOGLE_AUTH_START_FAILED_MESSAGE);
+        setIsGoogleLoading(false);
+      }
+    } catch {
+      setError(GOOGLE_AUTH_START_FAILED_MESSAGE);
+      setIsGoogleLoading(false);
+    }
+  }
+
+  return (
+    <div className="flex h-full min-h-screen items-center justify-center bg-background p-4">
+      <Card className="w-full max-w-md">
+        <CardHeader className="text-center">
+          <div className="mb-2 flex justify-center">
+            <Logo size="lg" />
+          </div>
+          <CardTitle className="text-lg font-semibold">Sign in to product</CardTitle>
+        </CardHeader>
+        <CardContent className="space-y-5">
+          {error && (
+            <div className="border border-red-200 bg-red-50 p-3 text-[12px] text-red-600 dark:border-red-900 dark:bg-red-950 dark:text-red-400">
+              {error}
+            </div>
+          )}
+
+          <form onSubmit={handleSubmit(onSubmit)} className="space-y-4">
+            <div>
+              <label className="mb-1 block text-[13px] font-medium">Email</label>
+              <Input
+                type="email"
+                placeholder="you@example.com"
+                className="h-8 text-[13px]"
+                {...register("email")}
+              />
+              {errors.email && (
+                <p className="mt-1 text-[11px] text-red-500">{errors.email.message}</p>
+              )}
+            </div>
+
+            <div>
+              <label className="mb-1 block text-[13px] font-medium">Password</label>
+              <div className="relative">
+                <Input
+                  type={showPassword ? "text" : "password"}
+                  placeholder="********"
+                  className="h-8 pr-10 text-[13px]"
+                  {...register("password")}
+                />
+                <button
+                  type="button"
+                  onClick={() => setShowPassword(!showPassword)}
+                  className="absolute right-3 top-1/2 -translate-y-1/2 text-muted-foreground hover:text-foreground"
+                >
+                  {showPassword ? <EyeOff size={16} /> : <Eye size={16} />}
+                </button>
+              </div>
+              {errors.password && (
+                <p className="mt-1 text-[11px] text-red-500">{errors.password.message}</p>
+              )}
+            </div>
+
+            <Button type="submit" size="sm" className="h-8 w-full text-[13px]" disabled={isLoading}>
+              {isLoading ? "Signing in..." : "Sign in"}
+            </Button>
+          </form>
+
+          <div className="relative">
+            <div className="absolute inset-0 flex items-center">
+              <span className="w-full border-t" />
+            </div>
+            <div className="relative flex justify-center text-[11px] uppercase">
+              <span className="bg-card px-2 text-muted-foreground">Or continue with</span>
+            </div>
+          </div>
+
+          <Button
+            variant="outline"
+            size="sm"
+            className="h-8 w-full text-[13px]"
+            onClick={handleGoogleSignIn}
+            disabled={isGoogleLoading}
+          >
+            {isGoogleLoading ? "Redirecting..." : "Google"}
+          </Button>
+
+          <p className="text-center text-[12px] text-muted-foreground">
+            Don&apos;t have an account?{" "}
+            <Link href="/auth/sign-up" className="font-medium text-primary hover:underline">
+              Sign up
+            </Link>
+          </p>
+        </CardContent>
+      </Card>
+    </div>
+  );
+}
+
+export function SignInClient({ googleAuthConfigured }: SignInContentProps) {
+  return (
+    <Suspense>
+      <SignInContent googleAuthConfigured={googleAuthConfigured} />
+    </Suspense>
+  );
+}
