+++ frontend/ui/src/lib/auth-errors.ts
@@ -1,0 +1,4 @@
+export const GOOGLE_AUTH_NOT_CONFIGURED_MESSAGE =
+  "Google sign-in is not configured for this environment. Set AUTH_GOOGLE_CLIENT_ID and AUTH_GOOGLE_CLIENT_SECRET, then restart the app.";
+
+export const GOOGLE_AUTH_START_FAILED_MESSAGE = "Unable to start Google sign-in. Please try again.";
